workaround failure of get_url on python 2.7.6 hosts
golang.org download URLs do not serve the correct TLS certificate
unless TLS SNI is used. The TLS client in Python 2.7.6 doesn't
support TLS SNI and so reports an certificate validation error.

Fortunately, the golang.org URL is actually a redirect to a
different URL which does export the correct certificate even
without SNI.

This change adjusts the specified download URL to take account of
which version of python is installed on the host by fallowing
the redirect with curl, defaulting to the original URL if
this process does not work for some reason.

Signed-off-by: Jon Seymour <jon@wildducktheories.com>